+++ lib/dsu/models/entry_group.rb
@@ -51,19 +51,6 @@ module Dsu
         def exists?(time:)
           Dsu::Services::EntryGroupReaderService.entry_group_file_exists?(time: time)
         end
-
-        # Loads the EntryGroup from the file system and returns an
-        # instantiated EntryGroup object.
-        def load(time: nil)
-          load_entry_group_file_for(time: time)
-        end
-
-        # This function returns a hash whose :time and :entries
-        # key values are hydrated with instantiated Time and Entry
-        # objects.
-        def load_entry_group_for(time:)
-          load_entry_group_file_for(time: time)
-        end
       end
 
       def valid_unique_entries
@@ -91,6 +78,10 @@ module Dsu
         self.class.delete!(time: time)
         self.entries = []
         self
+      end
+
+      def time_formatted
+        formatted_time(time: time)
       end
 
       def save!