+++ lib/dsu/support/times_sortable.rb
@@ -18,13 +18,11 @@ module Dsu
         end
       end
 
+      # TODO: Do we have something else we can use here?
       def times_for(times:)
         start_date = times.max
         return times unless start_date.monday? || start_date.on_weekend?
 
-        # (0..3).map { |num| start_date - num.days }
-        # (start_date..-start_date.friday?).map { |time| time }
-        # (0..3).map { |num| start_date - num.days if start_date.on_weekend? || start_date.monday? }
         # If the start_date is a weekend or a Monday, then we need to include
         # start_date along with all the dates up to and including the previous
         # Monday.