+++ spec/dsu/services/configuration_loader_service_spec.rb
@@ -43,7 +43,9 @@ RSpec.describe Dsu::Services::ConfigurationLoaderService do
           'editor' => 'editor',
           'entries_display_order' => 'entries_display_order',
           'entries_folder' => 'entries_folder',
-          'entries_file_name' => 'entries_file_name'
+          'entries_file_name' => 'entries_file_name',
+          'carry_over_entries_to_today' => 'carry_over_entries_to_today',
+          'include_all' => 'include_all'
         }
       end
       # rubocop:enable Style/StringHashKeys